3.0.0b7-1-2
Refactoring:

- Moved Argon2 parameter checking to seperate function
- Moved readFileBuffer and donwload creation to fileUtils.js
- Formatted passwordGenerator.js
- FileEncryptionController: Moved link logic to controller method, uses more vanilla JS
- Moved Main App Components to `app/main`
- Moved Password Generator components to `app/passwordGenerator`
- changed menu links to relative for new pages

Bug fixes:
- tried to fix duplicate icon when copying password from password-generator.html
- slot editor did not save the changes sometimes
