--- app/indexPasswordGenerator.js
@@ -1,3 +1,0 @@
-import { PasswordGenerator, PasswordGeneratorController } from '../app/passwordGenerator.js';
-const generator = new PasswordGenerator();
-new PasswordGeneratorController(generator);
