+++ app/main/algorithms/AESGCMStream/EncryptTransform.js
