+++ app/main/algorithms/AESGCMStream/StreamProcessor.js
