+++ app/main/controllers/EncryptionController.js
