+++ app/main/controllers/TextEncryptionController.js
