+++ app/main/helpers/ElementHandler.js
