+++ app/main/helpers/LaddaButtonHandler.js
