+++ app/main/helpers/ShowNotification.js
