+++ app/main/interfaces/IEncryptionAlgorithm.js
