+++ app/main/services/configManagement/ApplicationEncryptionManager.js
