+++ app/main/services/configManagement/SessionKeyManager.js
