+++ app/main/utils/dataEncode.js
