+++ app/main/utils/misc.js
