+++ index.html
@@ -190,31 +190,31 @@
                         </li>
 
                         <li role="none" class="menu-item">
-                            <a href="/pages/faq.html" role="menuitem" class="menu-link spinner-link">
+                            <a href="./pages/faq.html" role="menuitem" class="menu-link spinner-link">
                                 <span class="menu-icon" aria-hidden="true"><i class="mdi mdi-comment-question"></i></span>
                                 <span class="menu-text">FAQ</span>
                             </a>
                         </li>
                         <li role="none" class="menu-item">
-                            <a href="/pages/how-it-works.html" role="menuitem" class="menu-link spinner-link">
+                            <a href="./pages/how-it-works.html" role="menuitem" class="menu-link spinner-link">
                                 <span class="menu-icon" aria-hidden="true"><i class="mdi mdi-file-document"></i></span>
                                 <span class="menu-text">How it works</span>
                             </a>
                         </li>
                         <li role="none" class="menu-item">
-                            <a href="/pages/security-and-privacy.html" role="menuitem" class="menu-link spinner-link">
+                            <a href="./pages/security-and-privacy.html" role="menuitem" class="menu-link spinner-link">
                                 <span class="menu-icon" aria-hidden="true"><i class="mdi mdi-upload-lock"></i></span>
                                 <span class="menu-text">Security &amp; Privacy</span>
                             </a>
                         </li>
                         <li role="none" class="menu-item">
-                            <a href="/pages/developer-documentation-technical-information.html" role="menuitem" class="menu-link spinner-link">
+                            <a href="./pages/developer-documentation-technical-information.html" role="menuitem" class="menu-link spinner-link">
                                 <span class="menu-icon" aria-hidden="true"><i class="mdi mdi-source-branch"></i></span>
                                 <span class="menu-text">Developer Information</span>
                             </a>
                         </li>
                         <li role="none" class="menu-item">
-                            <a href="/pages/encryption-glossary.html" role="menuitem" class="menu-link spinner-link">
+                            <a href="./pages/encryption-glossary.html" role="menuitem" class="menu-link spinner-link">
                                 <span class="menu-icon" aria-hidden="true"><i class="mdi mdi-globe-model"></i></span>
                                 <span class="menu-text">Glossary</span>
                             </a>
@@ -222,7 +222,7 @@
 
                         <li role="presentation" class="menu-title">Tools</li>
                         <li role="none" class="menu-item">
-                            <a href="/pages/password-generator.html" role="menuitem" class="menu-link spinner-link">
+                            <a href="./pages/password-generator.html" role="menuitem" class="menu-link spinner-link">
                                 <span class="menu-icon" aria-hidden="true"><i class="mdi mdi-form-textbox-password"></i></span>
                                 <span class="menu-text">Password Generator</span>
                             </a>
@@ -1177,10 +1177,10 @@
                             </div>
                             <div class="col-md-6">
                                 <div class="d-flex footer-links flex-column flex-md-row gap-2 gap-md-4 justify-content-center justify-content-md-end">
-                                    <a href="/pages/about.html">About</a>
-                                    <a href="/pages/privacy-policy.html">Privacy Policy</a>
-                                    <a href="/pages/download-app.html">Download Project</a>
-                                    <a href="/pages/contact.html">Contact</a>
+                                    <a href="./pages/about.html">About</a>
+                                    <a href="./pages/privacy-policy.html">Privacy Policy</a>
+                                    <a href="./pages/download-app.html">Download Project</a>
+                                    <a href="./pages/contact.html">Contact</a>
                                 </div>
                             </div>
                         </div>
@@ -1208,7 +1208,7 @@
         <!-- PW strenght checker JS -->
         <script src="assets/libs/check-password-strength/umd.js" defer></script>
 
-        <script type="module" src="app/index.js" defer></script>
+        <script type="module" src="app/main/index.js" defer></script>
         <script src="app/insertVersion.js" defer></script>
     </body>
 </html>
