+++ pages/privacy-policy.html
@@ -63,7 +63,7 @@
                     <!--- Menu -->
                     <ul class="menu" role="menubar">
                         <li class="p-2 menu-item text-white">
-                            <a href="/index.html" class="spinner-link menu-link btn btn-primary">
+                            <a href="../index.html" class="spinner-link menu-link btn btn-primary">
                                 <span class="menu-icon text-white"><i class="mdi mdi-application-import" aria-hidden="true"></i></span>
                                 <span class="menu-text text-white">Open App</span>
                                 <span class="badge bg-light text-dark ms-auto">New</span>
@@ -76,31 +76,31 @@
                         </li>
 
                         <li class="menu-item">
-                            <a href="/pages/faq.html" class="menu-link spinner-link">
+                            <a href="./faq.html" class="menu-link spinner-link">
                                 <span class="menu-icon" aria-hidden="true"><i class="mdi mdi-comment-question"></i></span>
                                 <span class="menu-text">FAQ</span>
                             </a>
                         </li>
                         <li class="menu-item">
-                            <a href="/pages/how-it-works.html" class="menu-link spinner-link">
+                            <a href="./how-it-works.html" class="menu-link spinner-link">
                                 <span class="menu-icon" aria-hidden="true"><i class="mdi mdi-file-document"></i></span>
                                 <span class="menu-text">How it works</span>
                             </a>
                         </li>
                         <li class="menu-item">
-                            <a href="/pages/security-and-privacy.html" class="menu-link spinner-link">
+                            <a href="./security-and-privacy.html" class="menu-link spinner-link">
                                 <span class="menu-icon" aria-hidden="true"><i class="mdi mdi-upload-lock"></i></span>
                                 <span class="menu-text">Security & Privacy</span>
                             </a>
                         </li>
                         <li class="menu-item">
-                            <a href="/pages/developer-documentation-technical-information.html" class="menu-link spinner-link">
+                            <a href="./developer-documentation-technical-information.html" class="menu-link spinner-link">
                                 <span class="menu-icon" aria-hidden="true"><i class="mdi mdi-source-branch"></i></span>
                                 <span class="menu-text">Dev Information</span>
                             </a>
                         </li>
                         <li class="menu-item">
-                            <a href="/pages/encryption-glossary.html" class="menu-link spinner-link">
+                            <a href="./encryption-glossary.html" class="menu-link spinner-link">
                                 <span class="menu-icon" aria-hidden="true"><i class="mdi mdi-globe-model"></i></span>
                                 <span class="menu-text">Glossary</span>
                             </a>
@@ -108,7 +108,7 @@
 
                         <li class="menu-title">Tools</li>
                         <li class="menu-item">
-                            <a href="/pages/password-generator.html" class="menu-link spinner-link">
+                            <a href="./password-generator.html" class="menu-link spinner-link">
                                 <span class="menu-icon" aria-hidden="true"><i class="mdi mdi-form-textbox-password"></i></span>
                                 <span class="menu-text">Password Generator</span>
                             </a>
@@ -229,7 +229,7 @@
                                 <div class="page-title-right">
                                     <ol class="breadcrumb m-0">
                                         <li class="breadcrumb-item"><a href="https://encrypti0n.com">Encrypti0n.com</a></li>
-                                        <li class="breadcrumb-item active"><a href="/pages/privacy-policy.html">Privacy Policy</a></li>
+                                        <li class="breadcrumb-item active"><a href="./privacy-policy.html">Privacy Policy</a></li>
                                     </ol>
                                 </div>
                             </div>
@@ -335,10 +335,10 @@
                             </div>
                             <div class="col-md-6">
                                 <div class="d-flex footer-links flex-column flex-md-row gap-2 gap-md-4 justify-content-center justify-content-md-end">
-                                    <a href="/pages/about.html">About</a>
-                                    <a href="/pages/privacy-policy.html">Privacy Policy</a>
-                                    <a href="/pages/download-app.html">Download Project</a>
-                                    <a href="/pages/contact.html">Contact</a>
+                                    <a href="./about.html">About</a>
+                                    <a href="./privacy-policy.html">Privacy Policy</a>
+                                    <a href="./download-app.html">Download Project</a>
+                                    <a href="./contact.html">Contact</a>
                                 </div>
                             </div>
                         </div>
